+++ packages/google_maps_flutter/example/lib/heatmap.dart
@@ -1,4 +1,4 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -11,7 +11,7 @@ import 'page.dart';
 
 class HeatmapPage extends GoogleMapExampleAppPage {
   const HeatmapPage({Key? key})
-      : super(const Icon(Icons.map), 'Heatmaps', key: key);
+    : super(const Icon(Icons.map), 'Heatmaps', key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -127,13 +127,15 @@ class HeatmapBodyState extends State<HeatmapBody> {
                     Column(
                       children: <Widget>[
                         TextButton(
-                          onPressed:
-                              disabledPoints.isNotEmpty ? _addPoint : null,
+                          onPressed: disabledPoints.isNotEmpty
+                              ? _addPoint
+                              : null,
                           child: const Text('Add point'),
                         ),
                         TextButton(
-                          onPressed:
-                              enabledPoints.isNotEmpty ? _removePoint : null,
+                          onPressed: enabledPoints.isNotEmpty
+                              ? _removePoint
+                              : null,
                           child: const Text('Remove point'),
                         ),
                       ],