+++ packages/google_maps_flutter/example/lib/lite_mode.dart
@@ -1,4 +1,4 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -15,7 +15,7 @@ const CameraPosition _kInitialPosition = CameraPosition(
 
 class LiteModePage extends GoogleMapExampleAppPage {
   const LiteModePage({Key? key})
-      : super(const Icon(Icons.map), 'Lite mode', key: key);
+    : super(const Icon(Icons.map), 'Lite mode', key: key);
 
   @override
   Widget build(BuildContext context) {