+++ packages/google_maps_flutter/example/lib/main.dart
@@ -1,11 +1,15 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 import 'package:flutter/material.dart';
 
+import 'advanced_marker_icons.dart';
+import 'advanced_markers_clustering.dart';
 import 'animate_camera.dart';
 import 'clustering.dart';
+import 'collision_behavior.dart';
+import 'ground_overlay.dart';
 import 'heatmap.dart';
 import 'lite_mode.dart';
 import 'map_click.dart';
@@ -16,6 +20,7 @@ import 'marker_icons.dart';
 import 'move_camera.dart';
 import 'padding.dart';
 import 'page.dart';
+import 'place_advanced_marker.dart';
 import 'place_circle.dart';
 import 'place_marker.dart';
 import 'place_polygon.dart';
@@ -24,6 +29,10 @@ import 'scrolling_map.dart';
 import 'snapshot.dart';
 import 'tile_overlay.dart';
 
+/// Place your map ID here. Map ID is required for pages that use advanced
+/// markers.
+const String? _mapId = null;
+
 final List<GoogleMapExampleAppPage> _allPages = <GoogleMapExampleAppPage>[
   const MapUiPage(),
   const MapCoordinatesPage(),
@@ -31,7 +40,9 @@ final List<GoogleMapExampleAppPage> _allPages = <GoogleMapExampleAppPage>[
   const AnimateCameraPage(),
   const MoveCameraPage(),
   const PlaceMarkerPage(),
+  const PlaceAdvancedMarkerPage(mapId: _mapId),
   const MarkerIconsPage(),
+  const AdvancedMarkerIconsPage(mapId: _mapId),
   const ScrollingMapPage(),
   const PlacePolylinePage(),
   const PlacePolygonPage(),
@@ -40,9 +51,12 @@ final List<GoogleMapExampleAppPage> _allPages = <GoogleMapExampleAppPage>[
   const SnapshotPage(),
   const LiteModePage(),
   const TileOverlayPage(),
+  const GroundOverlayPage(),
   const ClusteringPage(),
+  const AdvancedMarkersClustering(mapId: _mapId),
   const MapIdPage(),
   const HeatmapPage(),
+  const AdvancedMarkerCollisionBehaviorPage(mapId: _mapId),
 ];
 
 /// MapsDemo is the Main Application.
@@ -53,8 +67,10 @@ class MapsDemo extends StatelessWidget {
   void _pushPage(BuildContext context, GoogleMapExampleAppPage page) {
     Navigator.of(context).push(
       MaterialPageRoute<void>(
-        builder: (_) =>
-            Scaffold(appBar: AppBar(title: Text(page.title)), body: page),
+        builder: (_) => Scaffold(
+          appBar: AppBar(title: Text(page.title)),
+          body: page,
+        ),
       ),
     );
   }