+++ packages/google_maps_flutter/example/lib/place_polyline.dart
@@ -1,4 +1,4 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -12,7 +12,7 @@ import 'page.dart';
 
 class PlacePolylinePage extends GoogleMapExampleAppPage {
   const PlacePolylinePage({Key? key})
-      : super(const Icon(Icons.linear_scale), 'Place polyline', key: key);
+    : super(const Icon(Icons.linear_scale), 'Place polyline', key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -109,11 +109,11 @@ class PlacePolylineBodyState extends State<PlacePolylineBody> {
       return;
     }
 
-    final String polylineIdVal = 'polyline_id_$_polylineIdCounter';
+    final polylineIdVal = 'polyline_id_$_polylineIdCounter';
     _polylineIdCounter++;
-    final PolylineId polylineId = PolylineId(polylineIdVal);
+    final polylineId = PolylineId(polylineIdVal);
 
-    final Polyline polyline = Polyline(
+    final polyline = Polyline(
       polylineId: polylineId,
       consumeTapEvents: true,
       color: Colors.orange,
@@ -306,7 +306,7 @@ class PlacePolylineBodyState extends State<PlacePolylineBody> {
   }
 
   List<LatLng> _createPoints() {
-    final List<LatLng> points = <LatLng>[];
+    final points = <LatLng>[];
     final double offset = _polylineIdCounter.ceilToDouble();
     points.add(_createLatLng(51.4816 + offset, -3.1791));
     points.add(_createLatLng(53.0430 + offset, -2.9925));